+++ CHANGELOG.md
@@ -7,11 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+The two largest efforts of this release are documentation and "streamer" refactor.
+
 Removed
 -------
 
-- `mpservice.streamer.Streamer.{drop_first_n, peek_random}`.
-
+- `mpservice.streamer.Streamer.{drop_first_n, peek_random, drop_if, keep_if}`, and corresponding classes
+  `Dropper`.
+- `mpservice.streamer.Streamer.drop_exceptions`.
 
 Changed
 -------
@@ -19,12 +22,17 @@ Changed
 - `mpservice.streamer.Streamer.transform`: parameter `concurrency` used to default to 1 (i.e. no concurrency), now defaults to higher numbers (i.e. with concurrency).
 - `mpservice.mpserver.{Sequential, Ensemble}` were renamed to `SequentialServlet` and `EnsembleServlet` respectively.
 - `mpservice.streamer.Streamer.drain`: return count of elements processed, instead of the tuple of element count and exception count.
+- `mpservice.streamer.Streamer.peek` was refactored.
+- `mpservice.streamer.Streamer.transform` was renamed to `parmap`.
 
 Added or enhanced
 -----------------
 
 - Enhanced documentation. Started to host generated doc on Read the Docs.
-- New class ``mpservice.mpserver.CpuAffinity``.
+- New class `mpservice.mpserver.CpuAffinity`.
+- New method on `mpservice.streamer.Streamer` and corresponding classes:
+  `filter` and `Filter`, `tail` and `Tailor`, `map` and `Mapper`, `groupby` and `Groupby`.
+- New method `mpservice.streamer.Streamer.filter_exceptions`.
 
 
 ## [0.11.7.post1] - 2022-10-21